requirements.txt

python-dotenv==1.0.0
chromadb==1.0.12
autogen-agentchat==0.6.1
google-generativeai==0.3.0
mistralai>=1.0.0
beautifulsoup4==4.13.4
requests==2.32.3
numpy>=1.21.0
scikit-learn==1.6.1
tiktoken==0.9.0
duckduckgo-search==3.5.0  


pyyaml==6.0.2  
kubernetes>=28.1.0  
pillow==11.2.1  
scipy==1.15.3  
wrapt==1.17.2  
protobuf==5.29.3  
sympy  


matplotlib>=3.5.0  
seaborn>=0.12.0  
pandas>=1.3.0  


